Start to change the structure of the plugin
