+++ plugins/id02/xpcs.py
@@ -1,0 +1,14 @@
+"""
+X-ray photon correlation spectroscopy plugin for ID02 
+
+"""
+
+__authors__ = ["Jérôme Kieffer"]
+__contact__ = "Jerome.Kieffer@ESRF.eu"
+__license__ = "MIT"
+__copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
+__date__ = "16/03/2020"
+__status__ = "development"
+__version__ = "0.0.1"
+
+"This is just a place-holder"
